% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS.R
\name{EBS}
\alias{EBS}
\title{accelerated empirical bayesian}
\usage{
EBS(data, conditions, gclus, sf, iter = 10, hyper, PP)
}
\arguments{
\item{data}{single cell expression matrix, row as genes column as cells}

\item{conditions}{partition of cells}

\item{gclus}{partition of genes}

\item{sf}{size factors}

\item{iter}{maximum iteration step of EM}

\item{hyper}{hyper parameters for beta distributions}

\item{different}{pattern of partitions}
}
\value{
posterior probability of mean expression pattern
}
\description{
accelerated empirical bayesian
}
