% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS_D.R
\name{EBS_D}
\alias{EBS_D}
\title{adjusted distance matrix by EBSeq}
\usage{
EBS_D(ccl, part, Y)
}
\arguments{
\item{ccl}{cluster of cells}

\item{part}{partition patterns}

\item{Y}{map pattern}
}
\value{
adjusted distance matrix by EBSeq
}
\description{
adjusted distance matrix by EBSeq
}
