% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBS_D_random.R
\name{EBS_D_random}
\alias{EBS_D_random}
\title{add random noise to distance matrix and generate new EBS distance}
\usage{
EBS_D_random(data, K, gcl, sz, D, hp, Posp, iter, lambda)
}
\arguments{
\item{data}{transcripts}

\item{K}{number of clusters}

\item{gcl}{gene cluster}

\item{sz}{size factor}

\item{D}{cell distance matrix}

\item{hp}{hyper parameters of EBSeq}

\item{Posp}{partition patterns}

\item{iter}{max number of iterations in EM algorithm in EBSeq}

\item{lambda}{parameter for exponential distributed noise}
}
\value{
new distance matrix
}
\description{
add random noise to distance matrix and generate new EBS distance
}
