% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cl.R
\name{g_cl}
\alias{g_cl}
\title{gene_level cluster}
\usage{
g_cl(data, ncores)
}
\arguments{
\item{data}{transcripts}

\item{ncores}{number of cores for parallel computing}
}
\value{
return a matrix whose row represent gene specific cluster
}
\description{
gene_level cluster
}
