% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRClus.R
\name{genRClus}
\alias{genRClus}
\title{determine number of clusters}
\usage{
genRClus(D, a, K, seed)
}
\arguments{
\item{D}{distance matrix of cells}

\item{a}{paramter for weights}

\item{K}{number of subtypes}

\item{seed}{seed for randomness}
}
\value{
random generated clustering of cells
}
\description{
determine number of clusters
}
