% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isref.R
\name{isref}
\alias{isref}
\title{check refinement relation between two clusters}
\usage{
isref(x, y)
}
\arguments{
\item{x}{a cluster}

\item{y}{a cluster}
}
\value{
whether x refines y
}
\description{
check refinement relation between two clusters
}
