% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listsize.R
\name{listsize}
\alias{listsize}
\title{number of DD genes under FDR control}
\usage{
listsize(pDD, FDR = 0.01)
}
\arguments{
\item{pDD}{estimated probability of being DD}

\item{FDR}{fdr to be controlled}
}
\value{
number of positive genes
}
\description{
number of DD genes under FDR control
}
