% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{likelihood function for hyperparameters estimation}
\usage{
LL(param, x, d0)
}
\arguments{
\item{param}{parameters to be determined by MLE}

\item{d0}{rate parameter of prior of 1 / true distance}

\item{D}{distance matrix of cells}
}
\value{
return hyperparameteres a.
}
\description{
likelihood function for hyperparameters estimation
}
