% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsz.R
\name{lsz}
\alias{lsz}
\title{index of DD genes under FDR control}
\usage{
lsz(pDD, FDR = 0.01)
}
\arguments{
\item{pDD}{probability of genes being DD}

\item{FDR}{fdr to be controlled}
}
\value{
index of positive genes
}
\description{
index of DD genes under FDR control
}
