% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP.R
\name{MCP}
\alias{MCP}
\title{modal-cluster for poisson gamma model}
\usage{
MCP(x, y = 1, z = c(1, 1))
}
\arguments{
\item{x}{an integer vector}

\item{y}{mass parameter}

\item{z}{shape and scale parameters for gamma distribution}
}
\value{
cluster of x
}
\description{
modal-cluster for poisson gamma model
}
\examples{
MCP(1:10)
}
