% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{MDD}
\alias{MDD}
\title{posterior of proportion change given mixture double dirichlet prior}
\usage{
MDD(z1, z2, pat, alpha1, alpha2)
}
\arguments{
\item{z1}{counts of each group in condition 1}

\item{z2}{counts of each group in condition 2}

\item{pat}{partition patterns}
}
\value{
posterior of proportion change
}
\description{
posterior of proportion change given mixture double dirichlet prior
}
