% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_D.R
\name{new_D}
\alias{new_D}
\title{generating new distance matrix between cells after EBSeq analysis}
\usage{
new_D(X, part, Y)
}
\arguments{
\item{X}{cell clusters}

\item{part}{partition patterns, obtained through Posp function}

\item{Y}{index of map pattern}
}
\value{
distance matrix of cells
}
\description{
generating new distance matrix between cells after EBSeq analysis
}
