% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD.R
\name{PDD}
\alias{PDD}
\title{calculate posterior probabilities of a gene to be differential distributed}
\usage{
PDD(data, cd, ncores, D, random = T, norm = T, epi = 1, Upper = 1000,
  nrandom = 30, iter = 20, reltol = 0.001, stp1 = 1e-06, stp2 = 0.01,
  K = 0)
}
\arguments{
\item{data}{normalized preprocessed transcripts}

\item{cd}{conditions label}

\item{ncores}{number of cores for parallel computing}

\item{D}{distance matrix of cells or cluster of cells or a given clustering}

\item{random}{boolean indicator of whether randomzation has been been implemented on distance matrix}

\item{epi}{tol for change of validity score in determining number of clusters}

\item{Upper}{bound for hyper parameters optimization}

\item{nrandom}{number of random generated distance matrix}

\item{iter}{max number of iterations for EM}

\item{reltol}{relative tolerance for optim on weighting paramters}

\item{K}{number of subtypes, could be user specified or determined internally}
}
\value{
posterior probabilities of a gene to be differential distributed
}
\description{
calculate posterior probabilities of a gene to be differential distributed
}
\examples{
data(sim_dat)
data_counts = assays(sim_dat)$count
cd = colData(sim_dat)$conditions
D_c = cal_D(data_counts,4)
pDD = PDD(data_counts,cd,2,D_c)
}
