% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_random.R
\name{PDD_random}
\alias{PDD_random}
\title{calculate PDD when add random noise in distance matrix}
\usage{
PDD_random(data, cd, K, D, a, sz, hp, Posp, iter, REF, seed)
}
\arguments{
\item{data}{normalized preprocessed transcripts}

\item{cd}{condition label}

\item{K}{number of subgroups}

\item{D}{distance matrix of cells}

\item{a}{shape param for weights}

\item{sz}{size factors}

\item{hp}{hyper parameters for EBSeq}

\item{Posp}{parition patterns}

\item{iter}{max number of iterations for EM in EBSeq}

\item{REF}{refinement relation matrix}

\item{b}{rate param for weights}

\item{random}{seed}
}
\value{
posterior probabilities under random distance matrix
}
\description{
calculate PDD when add random noise in distance matrix
}
