% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{preprocess data, remove non expressed genes}
\usage{
prep(data)
}
\arguments{
\item{data}{normalized transcripts}
}
\value{
preprocessed transcripts
}
\description{
preprocess data, remove non expressed genes
}
