% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMLE.R
\name{rwMLE}
\alias{rwMLE}
\title{MLE for random weighting parameter}
\usage{
rwMLE(D, reltol)
}
\arguments{
\item{D, }{distance matrix of cells}

\item{reltol, }{tolerance of convergence}
}
\value{
MLE of random weighting parameter
}
\description{
MLE for random weighting parameter
}
