% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{scDDboost}
\format{An object of class \code{"list"}.}
\usage{
data(sim_dat)
}
\description{
simulated data for demonstration, data are mixture negative binomial distributed
}
\examples{
data(sim_dat)
}
\keyword{datasets}
