% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_D.R
\name{cal_D}
\alias{cal_D}
\title{calculate distance matrix}
\usage{
cal_D(data, ncores)
}
\arguments{
\item{data}{transcripts}

\item{ncores}{number of cores for parallel computing}
}
\value{
distance matrix
}
\description{
calculate distance matrix
}
\examples{
data(sim_dat)
data_counts = assays(sim_dat)$count
D_c = cal_D(data_counts,4)
}
