% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{likelihood function for hyperparameters estimation}
\usage{
LL(param, x)
}
\arguments{
\item{D}{distance matrix of cells}
}
\value{
return hyperparameteres a.
}
\description{
likelihood function for hyperparameters estimation
}
