% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD.R
\name{PDD}
\alias{PDD}
\title{calculate posterior probabilities of a gene to be differential distributed}
\usage{
PDD(data, cd, ncores, D, random = T, epi = 0.15, Upper = 1000,
  nrandom = 30, iter = 20)
}
\arguments{
\item{data}{normalized preprocessed transcripts}

\item{cd}{conditions label}

\item{ncores}{number of cores for parallel computing}

\item{D}{distance matrix of cells or cluster of cells or a given clustering}

\item{random}{boolean indicator of whether randomzation has been been implemented on distance matrix}

\item{epi}{tol for change of validity score in determining number of clusters}

\item{Upper}{bound for hyper parameters optimization}

\item{nrandom}{number of random generated distance matrix}

\item{iter}{max number of iterations for EM}
}
\value{
posterior probabilities of a gene to be differential distributed
}
\description{
calculate posterior probabilities of a gene to be differential distributed
}
